Add files via upload
Project: Understand the results of an A/B test run by an e-commerce website. The company developed a new web page in an attempt to try to increase the number of users who "convert," i.e., the number of users who decide to pay for the company's product. The objective was to work through this Jupyter notebook to help the company understand if the new web page should be implemented; the old page should be kept; or, if this potential solution should be continued longer before making a final decision.